requirements.txt
newsapi-python==0.2.7
requests==2.31.0
python-dotenv==1.0.0
pandas==2.1.1
langchain
openai
langchain-openai==0.1.10
beautifulsoup4==4.12.3
weasyprint==61.2
markdown==3.6
jinja2==3.1.3